hoist common MqttEntity properties
