+++ homeassistant/components/mqtt/device_tracker/schema_discovery.py
@@ -10,10 +10,7 @@ from homeassistant.const import (
     ATTR_GPS_ACCURACY,
     ATTR_LATITUDE,
     ATTR_LONGITUDE,
-    CONF_DEVICE,
-    CONF_ICON,
     CONF_NAME,
-    CONF_UNIQUE_ID,
     CONF_VALUE_TEMPLATE,
     STATE_HOME,
     STATE_NOT_HOME,
@@ -27,7 +24,7 @@ from ..const import CONF_QOS, CONF_STATE_TOPIC
 from ..debug_info import log_messages
 from ..mixins import (
     MQTT_AVAILABILITY_SCHEMA,
-    MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+    MQTT_ENTITY_BASE_SCHEMA,
     MQTT_JSON_ATTRS_SCHEMA,
     MqttEntity,
     async_setup_entry_helper,
@@ -38,15 +35,13 @@ CONF_PAYLOAD_NOT_HOME = "payload_not_home"
 CONF_SOURCE_TYPE = "source_type"
 
 PLATFORM_SCHEMA_DISCOVERY = (
-    mqtt.MQTT_RO_PLATFORM_SCHEMA.extend(
+    mqtt.MQTT_RO_PLATFORM_SCHEMA.extend(MQTT_ENTITY_BASE_SCHEMA.schema)
+    .extend(
         {
-            vol.Optional(CONF_DEVICE): MQTT_ENTITY_DEVICE_INFO_SCHEMA,
-            vol.Optional(CONF_ICON): cv.icon,
             vol.Optional(CONF_NAME): cv.string,
             vol.Optional(CONF_PAYLOAD_HOME, default=STATE_HOME): cv.string,
             vol.Optional(CONF_PAYLOAD_NOT_HOME, default=STATE_NOT_HOME): cv.string,
             vol.Optional(CONF_SOURCE_TYPE): vol.In(SOURCE_TYPES),
-            vol.Optional(CONF_UNIQUE_ID): cv.string,
         }
     )
     .extend(MQTT_AVAILABILITY_SCHEMA.schema)
@@ -87,8 +82,6 @@ class MqttDeviceTracker(MqttEntity, TrackerEntity):
 
     def _setup_from_config(self, config):
         """(Re)Setup the entity."""
-        self._config = config
-
         value_template = self._config.get(CONF_VALUE_TEMPLATE)
         if value_template is not None:
             value_template.hass = self.hass
@@ -126,11 +119,6 @@ class MqttDeviceTracker(MqttEntity, TrackerEntity):
         )
 
     @property
-    def icon(self):
-        """Return the icon of the device."""
-        return self._config.get(CONF_ICON)
-
-    @property
     def latitude(self):
         """Return latitude if provided in device_state_attributes or None."""
         if (
@@ -164,11 +152,6 @@ class MqttDeviceTracker(MqttEntity, TrackerEntity):
     def location_name(self):
         """Return a location name for the current location of the device."""
         return self._location_name
-
-    @property
-    def name(self):
-        """Return the name of the device tracker."""
-        return self._config.get(CONF_NAME)
 
     @property
     def source_type(self):
