+++ homeassistant/components/mqtt/lock.py
@@ -5,13 +5,7 @@ import voluptuous as vol
 
 from homeassistant.components import lock
 from homeassistant.components.lock import LockEntity
-from homeassistant.const import (
-    CONF_DEVICE,
-    CONF_NAME,
-    CONF_OPTIMISTIC,
-    CONF_UNIQUE_ID,
-    CONF_VALUE_TEMPLATE,
-)
+from homeassistant.const import CONF_NAME, CONF_OPTIMISTIC, CONF_VALUE_TEMPLATE
 from homeassistant.core import callback
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.reload import async_setup_reload_service
@@ -30,7 +24,7 @@ from .. import mqtt
 from .debug_info import log_messages
 from .mixins import (
     MQTT_AVAILABILITY_SCHEMA,
-    MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+    MQTT_ENTITY_BASE_SCHEMA,
     MQTT_JSON_ATTRS_SCHEMA,
     MqttEntity,
     async_setup_entry_helper,
@@ -50,9 +44,9 @@ DEFAULT_STATE_LOCKED = "LOCKED"
 DEFAULT_STATE_UNLOCKED = "UNLOCKED"
 
 PLATFORM_SCHEMA = (
-    mqtt.MQTT_RW_PLATFORM_SCHEMA.extend(
+    mqtt.MQTT_RW_PLATFORM_SCHEMA.extend(MQTT_ENTITY_BASE_SCHEMA.schema)
+    .extend(
         {
-            vol.Optional(CONF_DEVICE): MQTT_ENTITY_DEVICE_INFO_SCHEMA,
             vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
             vol.Optional(CONF_OPTIMISTIC, default=DEFAULT_OPTIMISTIC): cv.boolean,
             vol.Optional(CONF_PAYLOAD_LOCK, default=DEFAULT_PAYLOAD_LOCK): cv.string,
@@ -63,7 +57,6 @@ PLATFORM_SCHEMA = (
             vol.Optional(
                 CONF_STATE_UNLOCKED, default=DEFAULT_STATE_UNLOCKED
             ): cv.string,
-            vol.Optional(CONF_UNIQUE_ID): cv.string,
         }
     )
     .extend(MQTT_AVAILABILITY_SCHEMA.schema)
@@ -111,8 +104,6 @@ class MqttLock(MqttEntity, LockEntity):
 
     def _setup_from_config(self, config):
         """(Re)Setup the entity."""
-        self._config = config
-
         self._optimistic = config[CONF_OPTIMISTIC]
 
         value_template = self._config.get(CONF_VALUE_TEMPLATE)
@@ -152,11 +143,6 @@ class MqttLock(MqttEntity, LockEntity):
                     }
                 },
             )
-
-    @property
-    def name(self):
-        """Return the name of the lock."""
-        return self._config[CONF_NAME]
 
     @property
     def is_locked(self):
