+++ homeassistant/components/mqtt/number.py
@@ -6,13 +6,7 @@ import voluptuous as vol
 
 from homeassistant.components import number
 from homeassistant.components.number import NumberEntity
-from homeassistant.const import (
-    CONF_DEVICE,
-    CONF_ICON,
-    CONF_NAME,
-    CONF_OPTIMISTIC,
-    CONF_UNIQUE_ID,
-)
+from homeassistant.const import CONF_NAME, CONF_OPTIMISTIC
 from homeassistant.core import callback
 from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.reload import async_setup_reload_service
@@ -32,7 +26,7 @@ from .const import CONF_RETAIN
 from .debug_info import log_messages
 from .mixins import (
     MQTT_AVAILABILITY_SCHEMA,
-    MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+    MQTT_ENTITY_BASE_SCHEMA,
     MQTT_JSON_ATTRS_SCHEMA,
     MqttEntity,
     async_setup_entry_helper,
@@ -44,13 +38,11 @@ DEFAULT_NAME = "MQTT Number"
 DEFAULT_OPTIMISTIC = False
 
 PLATFORM_SCHEMA = (
-    mqtt.MQTT_RW_PLATFORM_SCHEMA.extend(
+    mqtt.MQTT_RW_PLATFORM_SCHEMA.extend(MQTT_ENTITY_BASE_SCHEMA.schema)
+    .extend(
         {
-            vol.Optional(CONF_DEVICE): MQTT_ENTITY_DEVICE_INFO_SCHEMA,
-            vol.Optional(CONF_ICON): cv.icon,
             vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
             vol.Optional(CONF_OPTIMISTIC, default=DEFAULT_OPTIMISTIC): cv.boolean,
-            vol.Optional(CONF_UNIQUE_ID): cv.string,
         }
     )
     .extend(MQTT_AVAILABILITY_SCHEMA.schema)
@@ -90,7 +82,6 @@ class MqttNumber(MqttEntity, NumberEntity, RestoreEntity):
 
         self._current_number = None
         self._optimistic = config.get(CONF_OPTIMISTIC)
-        self._unique_id = config.get(CONF_UNIQUE_ID)
 
         NumberEntity.__init__(self)
         MqttEntity.__init__(self, None, config, config_entry, discovery_data)
@@ -99,9 +90,6 @@ class MqttNumber(MqttEntity, NumberEntity, RestoreEntity):
     def config_schema():
         """Return the config schema."""
         return PLATFORM_SCHEMA
-
-    def _setup_from_config(self, config):
-        self._config = config
 
     async def _subscribe_topics(self):
         """(Re)Subscribe to topics."""
@@ -166,16 +154,6 @@ class MqttNumber(MqttEntity, NumberEntity, RestoreEntity):
         )
 
     @property
-    def name(self):
-        """Return the name of this number."""
-        return self._config[CONF_NAME]
-
-    @property
     def assumed_state(self):
         """Return true if we do optimistic updates."""
         return self._optimistic
-
-    @property
-    def icon(self):
-        """Return the icon."""
-        return self._config.get(CONF_ICON)
